Add non-memfd backed exec memory back and refactor
