+++ libreflect/configure.ac
@@ -13,13 +13,18 @@ AM_PROG_CC_C_O
 AC_PROG_LIBTOOL
 LT_INIT
 
+HAVE_MEMFD=0
+
 case $host_os in
 	*linux*)
 		HOST_OS=linux
+		HAVE_MEMFD=1
 		CPPFLAGS="$CPPFLAGS -D_DEFAULT_SOURCE -D_BSD_SOURCE -D_POSIX_SOURCE -D_GNU_SOURCE"
 esac
 
 AM_CONDITIONAL([HOST_LINUX], [test x$HOST_OS = xlinux])
+AC_SUBST([HAVE_MEMFD], [$HAVE_MEMFD]) dnl TODO: actually check
+
 AC_SUBST([HOST_OS], [$HOST_OS])
 AC_SUBST([HOST_CPU], [$host_cpu])
